+++ src/idunno.Security.Ssrf/ProxiedSsrfDelegatingHandler.cs
@@ -4,6 +4,7 @@
 using System.Diagnostics.Metrics;
 using System.Net;
 using System.Net.Security;
+
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;