+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### Changed
 
-* URIs containing user information, e.g. `https://user:pass@example.com`, are now considered unsafe by `IsUnsafeUri`. `IsUnsafe` and the handlers.
+* URIs containing user information, e.g. `https://user:pass@example.com`, are now considered unsafe by `IsUnsafeUri`, `IsUnsafe` and
+the handlers.
 
 ## 5.0.0 - 2026-04-27