+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## 5.2.0 - Unreleased
+
+### Added
+
+* Add Azure WireServer IP range.
+* Add AMT/Global-Unicast range.
+* Add AS112 ranges.
+* Add Deprecated Anycast Prefix for 6to4 Relay Routers.
+* Add IPv6 discard/dummy range.
+* Add Segment Routing (SRv6) SIDs range.
+* Add tests for the specific ranges [AntiSSRF](https://github.com/microsoft/AntiSSRF/blob/main/config/IPAddressRanges.json) uses.
+
+### Changed
+
+* Stubbed the DNS resolver for mixed.ipv6.ssrf.fail tests because some corporate DNS servers can be flakey with that host, so
+  the tests are now deterministic.
+
 ## 5.1.0 - 2026-04-30
 
 ### Changed